% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEanalysisPreprocessing.R
\name{DEanalysisPreprocessing}
\alias{DEanalysisPreprocessing}
\title{Automatic creation of a DESeq2 object for DE analysis.}
\usage{
DEanalysisPreprocessing(
  RawCounts,
  Column.gene,
  Group.position,
  Time.position,
  Individual.position
)
}
\arguments{
\item{RawCounts}{Data.frame with \eqn{N_g} rows and (\eqn{N_{s+k}}) columns,
where \eqn{N_g} is the number of genes,
\eqn{N_s} is the number of samples and
\eqn{k=1} if a column is used to specify gene names, or \eqn{k=0} otherwise.
If \eqn{k=1}, the position of the column containing gene names is given
by \code{Column.gene}.
The data.frame contains non negative integers giving gene expressions of
each gene in each sample.
Column names of the data.frame must describe each sample's information
(individual, biological condition and time) and have the structure described
in the section \code{Details}.}

\item{Column.gene}{Integer indicating the column where gene names are given.
Set \code{Column.gene=NULL} if there is no such column.}

\item{Group.position}{Integer indicating the position of group information
in the string of characters in each sample names (see \code{Details}).
Set \code{Group.position=NULL} if there is only one or no biological
information in the string of character in each sample name.}

\item{Time.position}{Integer indicating the position of time measurement
information in the string of characters in each sample names
(see \code{Details}).
Set \code{Time.position=NULL} if there is only one or no time measurement
information in the string of character in each sample name.}

\item{Individual.position}{Integer indicating the position of the name of
the individual (e.g patient, replicate, mouse, yeasts culture ...)
in the string of characters in each sample names (see \code{Details}).
The names of different individuals must be all different.}
}
\value{
The function returns the \code{DESeq2} object for DE analysis and
the raw counts data (a matrix of non-negative integers).
}
\description{
This function creates automatically a DESeq2 object from raw counts data
using the R function \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSetFromMatrix()}}.
}
\details{
The column names of \code{ExprData} must be a vector of strings
of characters containing
\itemize{
\item a string of characters (if \eqn{k=1}) which is the label of the column
containing gene names.
\item \eqn{N_s} sample names which must be strings of characters containing
at least : the name of the individual (e.g patient, mouse, yeasts culture),
its biological condition (if there is at least two) and
the time where data have been collected if there is at least two;
(must be either 't0', 'T0' or '0' for time 0,
't1', 'T1' or '1' for time 1, ...).
}

All these sample information must be separated by underscores
in the sample name. For instance 'CLL_P_t0_r1',
corresponds to the patient 'r1' belonging to the biological condition 'P'
and where data were collected at time 't0'.
I this example, 'CLL' describe the type of cells
(here chronic lymphocytic leukemia) and is not used in our analysis.

In the string of characters 'CLL_P_t0_r1',
'r1' is localized after the third underscore, so \code{Individual.position=4},
'P' is localized after the first underscore, so \code{Group.position=2} and
't0' is localized after the second underscore, so \code{Time.position=3}.
}
\examples{
Grp.test=rep(c("P", "NP"), each=27)
Tps.test=rep(paste0("t", 0:8), times=6)
Pat.test=rep(paste0("pcl", 1:6), each=9)

Name.sel.test=paste(Grp.test, Pat.test, Tps.test, sep="_")
Mat.test.1=data.frame(Gene.name=paste0("Name", 1:10),
                      matrix(sample(1:100,length(Name.sel.test)*10,
                                    replace=TRUE),
                             ncol=length(Name.sel.test), nrow=10))
colnames(Mat.test.1)=c("Gene.name", Name.sel.test)
##-------------------------------------------------------------------------#
DESeq2.info.test=DEanalysisPreprocessing(RawCounts=Mat.test.1,
                                         Column.gene=1,
                                         Group.position=1,
                                         Time.position=3,
                                         Individual.position=2)
print(DESeq2.info.test)
}
\seealso{
The \code{\link[=DEanalysisPreprocessing]{DEanalysisPreprocessing()}} function
\itemize{
\item is used by the following functions of our package : \code{\link[=DATAnormalization]{DATAnormalization()}},
\code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}}.
\item calls the R function \code{\link[DESeq2:DESeqDataSet]{DESeq2::DESeqDataSetFromMatrix()}} in order to create
the DESeq2 object.
}
}
