% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEApreprocessing.R
\name{GSEApreprocessing}
\alias{GSEApreprocessing}
\title{GSEA preprocessing for official software and online tools.}
\usage{
GSEApreprocessing(
  Res.DE.analysis,
  ColumnsCriteria,
  Set.Operation,
  Rnk.files = TRUE,
  Save.files = FALSE
)
}
\arguments{
\item{Res.DE.analysis}{A list corresponding to the output from
\code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}} (see \code{Examples}).}

\item{ColumnsCriteria}{A vector of integers where each integer indicates a
column of \code{Res.DE.analysis$DE.results}.
These columns should either contain only binary values, or may contain other
numerical value, in which case extracted rows from
\code{Res.DE.analysis$DE.results} will be those with >0 values
(see \code{Details}).}

\item{Set.Operation}{A character. The user must choose between "union"
(default), "intersect", "setdiff" (see \code{Details}).}

\item{Rnk.files}{\code{TRUE} or \code{FALSE}. \code{TRUE} as default.
If \code{TRUE}, the rnk files generated by the function (used by the GSEA
software) will be saved if \code{Save.files=TRUE} and \code{path.result}
of \code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}} is not NULL.
Otherwise the rnk files will not be generated.}

\item{Save.files}{\code{TRUE} or \code{FALSE} or a Character.
If \code{Save.files=TRUE} and the \code{path.result} of \code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}}
is not NULL, all files will be saved in
"2_SupervisedAnalysis_\code{Name.folder.DE}/2-5_Enrichment_analysis_\code{Name.folder.DE}/
2-5-2_EnrichmentGO_software_preprocessing".
If \code{Save.files} is a character, it must be a path and all files
will be saved in the sub-folder "EnrichmentGO_software_preprocessing".
Otherwise, the different files will not be saved.}
}
\value{
The function returns
\itemize{
\item A vector of character containing gene names specified by
\code{ColumnsCriteria} and \code{Set.Operation}.
\item A vector of character containing all gene names
\item And, in case where \code{Save.files=TRUE} and the \code{path.result}
of \code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}} is not NULL, specific files designed to be used
as input for the following online tools and software :
\itemize{
\item GSEA : \url{https://www.gsea-msigdb.org/gsea/index.jsp}
\item DAVID : \url{https://david.ncifcrf.gov/tools.jsp}
\item WebGestalt : \url{http://www.webgestalt.org}
\item gProfiler : \url{https://biit.cs.ut.ee/gprofiler/gost}
\item Panther : \url{http://www.pantherdb.org}
\item ShinyGO : \url{http://bioinformatics.sdstate.edu/go/}
\item Enrichr : \url{https://maayanlab.cloud/Enrichr/}
\item GOrilla : \url{http://cbl-gorilla.cs.technion.ac.il}.
}
}
}
\description{
The function returns, from the output of \code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}},
specific files designed to be used as input for several online online tools
and software given in the section \code{Value}.
}
\details{
\itemize{
\item If \code{Set.Operation="union"} then the rows (so genes) extracted from
\code{Res.DE.analysis$DE.results} are those such that the sum of
the selected columns in \code{Res.DE.analysis} is >0.
For example, the rows extracted from \code{Res.DE.analysis$DE.results}
will be those DE at least at one time.
\item If \code{Set.Operation="intersect"} then the rows extracted from
\code{Res.DE.analysis$DE.results} are those such that the product of
the selected columns in \code{Res.DE.analysis$DE.results} is >0.
For example, the rows extracted from \code{Res.DE.analysis$DE.results}
will be those DE at all time ti (except the reference time t0).
\item If \code{Set.Operation="setdiff"} then the rows extracted from data are
those such that only one element of the selected columns in
\code{Res.DE.analysis$DE.results} is >0.
For example, the rows extracted from \code{Res.DE.analysis$DE.results}
will be those DE at only one time ti (except the reference time t0).
}
}
\examples{
data(Results_DEanalysis_sub500)
## Results of DEanalysisGlobal() with the dataset of Antoszewski
res.all<-Results_DEanalysis_sub500$DE_Antoszewski2022_MOUSEsub500
##
resGp<-GSEApreprocessing(Res.DE.analysis=res.all,
                         ColumnsCriteria=2,
                         Set.Operation="union",
                         Rnk.files=TRUE,
                         Save.files=FALSE)
##
##-------------------------------------------------------------------------#
## The results res.all of DEanalysisGlobal with the dataset Antoszewski2022
## data(RawCounts_Antoszewski2022_MOUSEsub500)
## res.all<-DEanalysisGlobal(RawCounts=RawCounts_Antoszewski2022_MOUSEsub500,
##                           Column.gene=1, Group.position=1,
##                           Time.position=NULL, Individual.position=2,
##                           pval.min=0.05, log.FC.min=1,
##                           LRT.supp.info=FALSE,
##                           path.result=NULL, Name.folder.DE=NULL)
}
