% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCPCanalysis.R
\name{HCPCanalysis}
\alias{HCPCanalysis}
\title{Hierarchical clustering analysis with HCPC (Main function)}
\usage{
HCPCanalysis(
  ExprData,
  Column.gene,
  Group.position,
  Time.position,
  Individual.position,
  sample.deletion,
  Supp.del.sample = FALSE,
  gene.deletion,
  Plot.HCPC = TRUE,
  Color.Group = NULL,
  Phi,
  Theta,
  Cex.point,
  epsilon,
  Cex.label,
  D3.mouvement = FALSE,
  path.result = NULL,
  Name.folder.hcpc = NULL
)
}
\arguments{
\item{ExprData}{Data.frame with \eqn{N_g} rows and (\eqn{N_{s+k}}) columns,
where \eqn{N_g} is the number of genes,
\eqn{N_s} is the number of samples and
\eqn{k=1} if a column is used to specify gene names, or \eqn{k=0} otherwise.
If \eqn{k=1}, the position of the column containing gene names is given
by \code{Column.gene}.
The data.frame contains numeric values giving gene expressions of each gene
in each sample.
Gene expressions can be raw counts or normalized raw counts.
Column names of the data.frame must describe each sample's information
(individual, biological condition and time) and have the structure described
in the section \code{Details}.}

\item{Column.gene}{Integer indicating the column where gene names are given.
Set \code{Column.gene=NULL} if there is no such column.}

\item{Group.position}{Integer indicating the position of group information
in the string of characters in each sample names (see \code{Details}).
Set \code{Group.position=NULL} if there is only one or no biological
information in the string of character in each sample name.}

\item{Time.position}{Integer indicating the position of time measurement
information in the string of characters in each sample names
(see \code{Details}).
Set \code{Time.position=NULL} if there is only one or no time measurement
information in the string of character in each sample name.}

\item{Individual.position}{Integer indicating the position of the name of
the individual (e.g patient, replicate, mouse, yeasts culture ...)
in the string of characters in each sample names (see \code{Details}).
The names of different individuals must be all different.
Furthermore, if individual names are just numbers, they will be transform in
a vector of class "character" by \code{\link[=CharacterNumbers]{CharacterNumbers()}} and
a "r" will be added to each individual name ("r" for replicate).}

\item{sample.deletion}{\code{NULL} or a vector of characters or a
vector of integers.
If \code{sample.deletion} is a vector of characters, all samples with names
in \code{sample.deletion} will not be used in the construction of the PCA.
If \code{sample.deletion} is a vector of integers, all the corresponding
column numbers of \code{ExprData} will not be used in the construction
of the PCA.
If \code{sample.deletion=NULL} all samples will be used in the construction
of the PCA.}

\item{Supp.del.sample}{\code{TRUE} or \code{FALSE}.
If \code{FALSE}, the samples selected with \code{sample.deletion}
will be deleted.
If \code{TRUE}, the samples selected with \code{sample.deletion}
will be plotted.
These individuals are called supplementary individuals in
\code{\link[FactoMineR:PCA]{FactoMineR::PCA()}}.}

\item{gene.deletion}{\code{NULL} or a vector of characters or a vector
of integers.
If \code{gene.deletion} is a vector of characters, all genes with names in
\code{gene.deletion} will be deleted from \code{ExprData}.
If \code{gene.deletion} is a vector of integers, all the corresponding
row numbers of \code{ExprData} will be deleted.
If \code{gene.deletion=NULL} all genes of \code{ExprData} will be used
in the construction of the PCA.}

\item{Plot.HCPC}{\code{TRUE} or \code{FALSE}. \code{TRUE} as default.
If \code{TRUE}, all graphs will be plotted.
Otherwise no graph will be plotted.}

\item{Color.Group}{NULL or a data.frame with \eqn{N_{bc}} rows and
two columns where \eqn{N_{bc}} is the number of biological conditions.
If \code{Color.Group} is a data.frame, the first column must contain
the name of each biological condition and the second column must contain
the colors associated to each biological condition.
If \code{Color.Group=NULL}, the function will automatically attribute
a color for each biological condition.
If samples belong to different time points only,
\code{Color.Group} will not be used.}

\item{Phi}{Angle defining the colatitude direction for the 3D PCA plot
(see \code{Details} in \code{\link[graphics:persp]{graphics::persp()}}).}

\item{Theta}{Angle defining the azimuthal direction for the 3D PCA plot
(see \code{Details} in \code{\link[graphics:persp]{graphics::persp()}}).}

\item{Cex.point}{Non negative numeric value giving the size of points
in all PCA plots which are not automatically plotted by \code{\link[FactoMineR:PCA]{FactoMineR::PCA()}}.}

\item{epsilon}{Non negative numeric value giving the length between points
and their labels in all PCA plots which are not automatically plotted
by \code{\link[FactoMineR:PCA]{FactoMineR::PCA()}}.}

\item{Cex.label}{Non negative numeric value giving the size of the labels
associated to each point of the all PCA graphs which are not automatically
plotted by \code{\link[FactoMineR:PCA]{FactoMineR::PCA()}}.}

\item{D3.mouvement}{\code{TRUE} or \code{FALSE}.
If \code{TRUE}, the 3D PCA plots will also be plotted in a rgl window
(see \code{\link[plot3Drgl:plotrgl]{plot3Drgl::plotrgl()}}) allowing to interactively rotate and zoom.}

\item{path.result}{Character or \code{NULL}. Path to save all results.
If \code{path.result} contains a sub folder entitled
"1_UnsupervisedAnalysis_\code{Name.folder.hcpc}" and a sub sub folder,
"1-3_HCPCanalysis_\code{Name.folder.hcpc}"
all results will be saved in the sub folder
"1_UnsupervisedAnalysis_\code{Name.folder.hcpc}/1-3_HCPCanalysis_\code{Name.folder.hcpc}".
Otherwise, a sub folder entitled
"1_UnsupervisedAnalysis_\code{Name.folder.hcpc}" and/or a sub sub folder
"1-3_HCPCanalysis_\code{Name.folder.hcpc}"
will be created in \code{path.result} and all results will be saved in
"1_UnsupervisedAnalysis_\code{Name.folder.hcpc}/1-3_HCPCanalysis_\code{Name.folder.hcpc}".
If \code{NULL}, the results will not be saved in a folder.
\code{NULL} as default.}

\item{Name.folder.hcpc}{Character or \code{NULL}.
If \code{Name.folder.hcpc} is a character, the folder and sub folder names
which will contain the PCA graphs will respectively be
"1_UnsupervisedAnalysis_\code{Name.folder.hcpc}"
and "1-3_HCPCanalysis_\code{Name.folder.hcpc}".
Otherwise, the folder and sub folder names will respectively be
"1_UnsupervisedAnalysis" and "1-3_HCPCanalysis".}
}
\value{
The function returns
\itemize{
\item a dendrogram (also called hierarchical tree) using the function
\code{\link[factoextra:fviz_dend]{factoextra::fviz_dend()}}
\item one 2D PCA and two 3D PCA produced by the function \code{\link[=PCArealization]{PCArealization()}}
where samples are colored with different colors for different clusters.
The two 3D PCA graphs are identical but one of them will be opened
in a rgl window (see \code{\link[plot3Drgl:plotrgl]{plot3Drgl::plotrgl()}}) allowing to interactively
rotate and zoom.
The interactive 3D graph will be plotted only if \code{D3.mouvement=TRUE}.
\item A graph indicating for each sample, its cluster and
the time and/or biological condition associated to the sample.
\item the outputs of \code{\link[FactoMineR:HCPC]{FactoMineR::HCPC()}}.
}
}
\description{
The functions performs a hierarchical clustering on results
from a factor analysis with the R function \code{\link[FactoMineR:HCPC]{FactoMineR::HCPC()}}.
}
\details{
The column names of \code{ExprData} must be a vector of strings
of characters containing
\itemize{
\item a string of characters (if \eqn{k=1}) which is the label of the column
containing gene names.
\item \eqn{N_s} sample names which must be strings of characters containing
at least : the name of the individual (e.g patient, mouse, yeasts culture),
its biological condition (if there is at least two) and
the time where data have been collected if there is at least two;
(must be either 't0', 'T0' or '0' for time 0,
't1', 'T1' or '1' for time 1, ...).
}

All these sample information must be separated by underscores
in the sample name. For instance 'CLL_P_t0_r1',
corresponds to the patient 'r1' belonging to the biological condition 'P'
and where data were collected at time 't0'.
I this example, 'CLL' describe the type of cells
(here chronic lymphocytic leukemia) and is not used in our analysis.

In the string of characters 'CLL_P_t0_r1',
'r1' is localized after the third underscore,
so \code{Individual.position=4},
'P' is localized after the first underscore, so \code{Group.position=2} and
't0' is localized after the second underscore, so \code{Time.position=3}.

The number of clusters is automatically selected by \code{\link[FactoMineR:HCPC]{FactoMineR::HCPC()}} and
is described in the section \code{Details} of \code{\link[FactoMineR:HCPC]{FactoMineR::HCPC()}}.
}
\examples{
res.sim.count<-RawCountsSimulation(Nb.Group=2, Nb.Time=3, Nb.per.GT=2,
                                   Nb.Gene=20)
##-------------------------------------------------------------------------#
Res.hcpc.analysis<-HCPCanalysis(ExprData=res.sim.count$Sim.dat,
                                Column.gene=1,
                                Group.position=1,
                                Time.position=2,
                                Individual.position=3,
                                sample.deletion=NULL,
                                Supp.del.sample=FALSE,
                                gene.deletion=NULL,
                                Plot.HCPC=TRUE,
                                Color.Group=NULL,
                                Phi=25, Theta=140,
                                Cex.point=1, Cex.label=0.6, epsilon=0.4,
                                D3.mouvement=FALSE,
                                path.result=NULL,
                                Name.folder.hcpc=NULL)
}
\seealso{
The function calls the functions \code{\link[=PCArealization]{PCArealization()}}
and \code{\link[FactoMineR:HCPC]{FactoMineR::HCPC()}}.
The function \code{\link[FactoMineR:HCPC]{FactoMineR::HCPC()}} will take as input the output
of \code{\link[=PCArealization]{PCArealization()}}.
}
