% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFUZZanalysis.R
\name{MFUZZanalysis}
\alias{MFUZZanalysis}
\title{Clustering of temporal patterns (Main function).}
\usage{
MFUZZanalysis(
  ExprData,
  Column.gene,
  Group.position,
  Time.position,
  Individual.position,
  DataNumberCluster,
  Method = "hcpc",
  Max.clust = 10,
  Membership,
  Min.std = 0.1,
  Plot.Mfuzz = TRUE,
  path.result = NULL,
  Name.folder.mfuzz = NULL
)
}
\arguments{
\item{ExprData}{Data.frame with \eqn{N_g} rows and (\eqn{N_{s+k}}) columns,
where \eqn{N_g} is the number of genes,
\eqn{N_s} is the number of samples and
\eqn{k=1} if a column is used to specify gene names, or \eqn{k=0} otherwise.
If \eqn{k=1}, the position of the column containing gene names is given
by \code{Column.gene}.
The data.frame contains numeric values giving gene expressions of each gene
in each sample.
Gene expressions can be raw counts or normalized raw counts.
Column names of the data.frame must describe each sample's information
(individual, biological condition and time) and have the structure described
in the section \code{Details}.}

\item{Column.gene}{Integer indicating the column where gene names are given.
Set \code{Column.gene=NULL} if there is no such column.}

\item{Group.position}{Integer indicating the position of group information
in the string of characters in each sample names (see \code{Details}).
Set \code{Group.position=NULL} if there is only one or no biological
information in the string of character in each sample name.}

\item{Time.position}{Integer indicating the position of time measurement
information in the string of characters in each sample names
(see \code{Details}).
Set \code{Time.position=NULL} if there is only one or no time measurement
information in the string of character in each sample name.}

\item{Individual.position}{Integer indicating the position of the name of
the individual (e.g patient, replicate, mouse, yeasts culture ...)
in the string of characters in each sample names (see \code{Details}).
The names of different individuals must be all different.
Furthermore, if individual names are just numbers, they will be transform in
a vector of class "character" by \code{\link[=CharacterNumbers]{CharacterNumbers()}} and
a "r" will be added to each individual name ("r" for replicate).}

\item{DataNumberCluster}{Data.frame or \code{NULL}. \code{NULL} as default.
If \code{DataNumberCluster} is a data.frame where the first column contains
the name of the biological conditions and the second the number of cluster
selected for each biological condition.
If \code{DataNumberCluster=NULL}, a number of clusters will be automatically
computed for each biological condition (see \code{\link[=MFUZZclustersNumber]{MFUZZclustersNumber()}}).}

\item{Method}{"kmeans" or "hcpc". The method used for selecting the number
of cluster to be used for the temporal cluster analysis
(see \code{Details}).
Only used if \code{DataNumberCluster} is not \code{NULL}.}

\item{Max.clust}{Integer strictly superior to 1 indicating the maximum
number of clusters.
\code{Max.clust} will be used only if \code{DataNumberCluster=NULL}}

\item{Membership}{Numeric value between 0 and 1.
For each cluster, genes with membership values below the threshold
\code{Membership} will not be displayed.
The membership values correspond to the probability of gene to belong
to each cluster.}

\item{Min.std}{Numeric positive value.
All genes where their standard deviations are smaller than the threshold
\code{Min.std} will be excluded.}

\item{Plot.Mfuzz}{\code{TRUE} or \code{FALSE}. \code{TRUE} as default.
If \code{TRUE}, all graphs will be plotted.
Otherwise no graph will be plotted.}

\item{path.result}{Character or \code{NULL}. Path to save all results.
If \code{path.result} contains a sub folder entitled
"1_UnsupervisedAnalysis_\code{Name.folder.mfuzz}" and a sub sub folder,
"1-4_MFUZZanalysis_\code{Name.folder.mfuzz}"
all results will be saved in the sub folder
"1_UnsupervisedAnalysis_\code{Name.folder.mfuzz}/1-4_MFUZZanalysis_\code{Name.folder.mfuzz}.
Otherwise, a sub folder entitled
"1_UnsupervisedAnalysis_\code{Name.folder.mfuzz}" and/or a sub sub folder
"1-4_MFUZZanalysis_\code{Name.folder.mfuzz}"
will be created in \code{path.result} and all results will be saved in
"1_UnsupervisedAnalysis_\code{Name.folder.mfuzz}/1-4_MFUZZanalysis_\code{Name.folder.mfuzz}".
If \code{NULL}, the results will not be saved in a folder.
\code{NULL} as default.}

\item{Name.folder.mfuzz}{Character or \code{NULL}.
If \code{Name.folder.mfuzz} is a character, the folder and sub folder names
which will contain the PCA graphs will respectively be
"1_UnsupervisedAnalysis_\code{Name.folder.mfuzz}" and
"1-4_MFUZZanalysis_\code{Name.folder.mfuzz}".
Otherwise, the folder and sub folder names will respectively be
"1_UnsupervisedAnalysis" and "1-4_MFUZZanalysis".}
}
\value{
The function returns
\itemize{
\item the final data used for the Mfuzz analysis (see \code{Details}).
\item the cluster associated to each gene.
\item plots generated by \code{\link[Mfuzz:mfuzz.plot2]{Mfuzz::mfuzz.plot2()}} for each biological condition.
}
}
\description{
The function performs a soft clustering of temporal patterns
based on the fuzzy c-means algorithm using the R package \code{Mfuzz}.
}
\details{
The column names of \code{ExprData} must be a vector of strings
of characters containing
\itemize{
\item a string of characters (if \eqn{k=1}) which is the label of the column
containing gene names.
\item \eqn{N_s} sample names which must be strings of characters containing
at least : the name of the individual (e.g patient, mouse, yeasts culture),
its biological condition (if there is at least two) and
the time where data have been collected if there is at least two;
(must be either 't0', 'T0' or '0' for time 0,
't1', 'T1' or '1' for time 1, ...).
}

All these sample information must be separated by underscores
in the sample name. For instance 'CLL_P_t0_r1',
corresponds to the patient 'r1' belonging to the biological condition 'P'
and where data were collected at time 't0'.
I this example, 'CLL' describe the type of cells
(here chronic lymphocytic leukemia) and is not used in our analysis.

In the string of characters 'CLL_P_t0_r1',
'r1' is localized after the third underscore,
so \code{Individual.position=4},
'P' is localized after the first underscore, so \code{Group.position=2} and
't0' is localized after the second underscore, so \code{Time.position=3}.

The \code{Mfuzz} package works with datasets where rows correspond to genes
and columns correspond to times.
If \code{ExprData} contains several replicates per time,
the algorithm computes the mean of replicates for each gene
before using \code{\link[Mfuzz:mfuzz]{Mfuzz::mfuzz()}}.
When there are several biological conditions, the algorithm realizes
the \code{\link[Mfuzz:mfuzz]{Mfuzz::mfuzz()}} analysis for each biological condition.
}
\examples{
# Data simulation
data.clust.sim<-matrix(rnorm(12*10*3,
                             mean=rep(c(rep(c(1,6,9,4,3,1,6.5,0.7,10),
                                            times=2),
                                        rep(c(2,3.6,3.7,5,7.9,8,7.5,3.5,3.4),
                                            times=2)),
                             each=10), sd=0.2),
                       nrow=30, ncol=12)
#
colnames(data.clust.sim)<-c("G1_t0_r1","G1_t1_r1","G1_t2_r1",
                            "G1_t0_r2","G1_t1_r2","G1_t2_r2",
                            "G2_t0_r3","G2_t1_r3","G2_t2_r3",
                            "G2_t0_r4","G2_t1_r4","G2_t2_r4")
##-------------------------------------------------------------------------#
# Plot the temporal expression of each individual
graphics::matplot(t(rbind(data.clust.sim[,1:3], data.clust.sim[,4:6],
                          data.clust.sim[,7:9], data.clust.sim[,10:12])),
                          type=c("b"), pch=19, col=rep(c("black","red"),
                          each=6*10), xlab="Time", ylab="Gene expression")
##-------------------------------------------------------------------------#
MFUZZanalysis(ExprData=data.clust.sim,
              Column.gene=NULL,
              Group.position=1,
              Time.position=2,
              Individual.position=3,
              DataNumberCluster=NULL,
              Membership=0.5,
              Min.std=0.1,
              Plot.Mfuzz=TRUE,
              path.result=NULL)
}
\seealso{
The function uses the function \code{\link[=MFUZZclustersNumber]{MFUZZclustersNumber()}} in order to
compute the optimal number of cluster for each biological condition
with the kmeans method.
}
