% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEplotHeatmaps.R
\name{DEplotHeatmaps}
\alias{DEplotHeatmaps}
\title{Heatmaps of DE genes}
\usage{
DEplotHeatmaps(
  Res.DE.analysis,
  ColumnsCriteria = 2,
  Set.Operation = "union",
  NbGene.analysis = 20,
  Color.Group = NULL,
  SizeLabelRows = 5,
  SizeLabelCols = 5,
  Display.plots = TRUE,
  Save.plots = FALSE
)
}
\arguments{
\item{Res.DE.analysis}{A list. Output from \code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}}.}

\item{ColumnsCriteria}{A vector of integers where each integer indicates
column of the output \code{DE.results} from \code{Res.DE.analysis}.
Columns should contained binary values, but if there are other
integers or other values, rows selected are those which values are >0
(see \code{Details}).}

\item{Set.Operation}{A character.
The user must choose between "union" (default), "intersect", "setdiff"
(see \code{Details}).}

\item{NbGene.analysis}{An integer or \code{NULL}.
If it is an integer, the heatmaps will be plotted with the
\code{NbGene.analysis} genes which have the highest sum of absolute
log2 fold change, among the DE genes selected using \code{ColumnsCriteria}
and \code{Set.Operation}.
If \code{NULL}, all the DE selected genes will be used for both heatmaps.}

\item{Color.Group}{\code{NULL} or a data.frame with \eqn{N_{bc}} rows and
two columns where \eqn{N_{bc}} is the number of biological conditions.
If \code{Color.Group} is a data.frame, the first column must contain
the name of each biological condition and the second column must contain
the colors associated to each biological condition.
If \code{Color.Group=NULL}, the function will automatically attribute
a color for each biological condition.
If samples belong to different time points only,
\code{Color.Group} will not be used.}

\item{SizeLabelRows}{Numeric >0.
Size of the labels for the genes in the heatmaps.}

\item{SizeLabelCols}{Numeric >0.
Size of the labels for the samples in the heatmaps.}

\item{Display.plots}{\code{TRUE} or \code{FALSE}. \code{TRUE} as default.
If \code{TRUE}, all graphs will be plotted.
Otherwise no graph will be plotted.}

\item{Save.plots}{TRUE or FALSE or a Character.
If \code{Save.plots=FALSE}, the different files will not be saved.
If \code{Save.plots=TRUE} and the \code{path.result} of \code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}}
is not NULL, all files will be saved in
"2_SupervisedAnalysis_\code{Name.folder.DE}/
2-4_Supplementary_Plots_\code{Name.folder.DE}/Plots_Heatmaps".
If \code{Save.plots} is a character, it must be a path and
all files will be saved in the sub-folder "Plots_Heatmaps".}
}
\value{
The function returns two heatmaps:
one heatmap of gene expressions between samples and selected genes;
and a correlation heatmap between samples.
}
\description{
The function returns two heatmaps:
one heatmap of gene expressions between samples and selected genes and
a correlation heatmap between samples
from the output of \code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}}.
}
\details{
\itemize{
\item If \code{Set.Operation="union"} then the rows extracted from \code{Data}
are those such that the sum of the selected columns in \code{Res.DE.analysis}
is >0.
For example, if \code{Res.DE.analysis} is the outputs from
\code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}}, the rows extracted from \code{Data} will be those DE
at least at one time.
\item If \code{Set.Operation="intersect"} then the rows extracted from data are
those such that the product of the selected columns in \code{Res.DE.analysis}
is >0.
For example, if \code{Res.DE.analysis} is the outputs from
\code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}}, the rows extracted from \code{Data} will be those DE
at all time ti (except the reference time t0).
\item If \code{Set.Operation="setdiff"} then the rows extracted from data are
those such that only one element of the selected columns in
\code{Res.DE.analysis} is >0.
For example, if \code{Res.DE.analysis} is the outputs from
\code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}}, the rows extracted from \code{Data} will be those DE
at only one time ti (except the reference time t0).
}

The time in the first column of \code{Color.Time} must be either
't0', 'T0' or '0' for time 0,
't1', 'T1' or '1' for time 1, ...
}
\examples{
data(Results_DEanalysis_sub500)
# Results of DEanalysisGlobal() with the dataset of Antoszewski
res.all<-Results_DEanalysis_sub500$DE_Antoszewski2022_MOUSEsub500
#
resHeatmap<-DEplotHeatmaps(Res.DE.analysis=res.all,
                           ColumnsCriteria=3, ## Specific genes N1haT1ko
                           Set.Operation="union",
                           NbGene.analysis=20,
                           Color.Group=NULL,
                           SizeLabelRows=5,
                           SizeLabelCols=5,
                           Display.plots=TRUE,
                           Save.plots=FALSE)
#
#--------------------------------------------------------------------------#
## The results res.all of DEanalysisGlobal with the dataset Antoszewski2022
# data(RawCounts_Antoszewski2022_MOUSEsub500)
# res.all<-DEanalysisGlobal(RawCounts=RawCounts_Antoszewski2022_MOUSEsub500,
#                           Column.gene=1, Group.position=1,
#                           Time.position=NULL, Individual.position=2,
#                           pval.min=0.05, log.FC.min=1,LRT.supp.info=FALSE,
#                           path.result=NULL, Name.folder.DE=NULL)
}
\seealso{
The function calls the function \code{\link[ComplexHeatmap:Heatmap]{ComplexHeatmap::Heatmap()}}
in order to plot the Heatmaps.
}
