% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_DEanalysis_sub500.R
\docType{data}
\name{Results_DEanalysis_sub500}
\alias{Results_DEanalysis_sub500}
\title{DE results of three dataset}
\format{
A list of 3 lists.
}
\usage{
data(Results_DEanalysis_sub500)
}
\description{
The list Results_DEanalysis_sub500 contains the results of
\code{\link[=DEanalysisGlobal]{DEanalysisGlobal()}} for each of the following raw counts :
RawCounts_Weger2021_MOUSEsub500, RawCounts_Leong2014_FISSIONsub500wt and
RawCounts_Schleiss2021_CLLsub500
}
\details{
Each list in Results_DEanalysis_sub500 contains the outputs
"Summary.Inputs", "DE.results" and "DESeq.dds" of DEanalysisGlobal()
for each of the following raw counts :
RawCounts_Weger2021_MOUSEsub500, RawCounts_Leong2014_FISSIONsub500wt and
RawCounts_Schleiss2021_CLLsub500
}
\examples{
data(Results_DEanalysis_sub500)
}
\keyword{datasets}
